requirements.txt
deta
fastapi
uvicorn
pyjwt
passlib[bcrypt]
email-validator==1.1.2
fastapi==0.62.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3